replace prop passing with connect between map-interaction components
