+++ web/js/containers/map-interactions/ol-vector-interactions.js
@@ -1,16 +1,24 @@
+import { connect } from 'react-redux';
+
 import {
   debounce as lodashDebounce,
   get as lodashGet,
   includes as lodashIncludes,
+  groupBy as lodashGroupBy,
 } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { transform } from 'ol/proj';
 import { isFromActiveCompareRegion } from '../../modules/compare/util';
 import { hasNonClickableVectorLayer } from '../../modules/layers/util';
+import vectorDialog from '../vector-dialog';
+import { onMapClickGetVectorFeatures } from '../../modules/vector-styles/util';
+import { openCustomContent, onClose } from '../../modules/modal/actions';
+import { selectVectorFeatures as selectVectorFeaturesActionCreator } from '../../modules/vector-styles/actions';
+import { changeCursor as changeCursorActionCreator } from '../../modules/map/actions';
+import { ACTIVATE_VECTOR_ALERT } from '../../modules/alerts/constants';
 
-
-export default class OlVectorInteractions extends React.Component {
+export class VectorInteractions extends React.Component {
   constructor(props) {
     super(props);
     this.mouseMove = lodashDebounce(this.mouseMove.bind(this), 8);
@@ -93,7 +101,7 @@ export default class OlVectorInteractions extends React.Component {
   }
 }
 
-OlVectorInteractions.propTypes = {
+VectorInteractions.propTypes = {
   changeCursor: PropTypes.func.isRequired,
   getDialogObject: PropTypes.func.isRequired,
   isShowingClick: PropTypes.bool.isRequired,
@@ -111,3 +119,94 @@ OlVectorInteractions.propTypes = {
   activeLayers: PropTypes.array,
   activateVectorAlert: PropTypes.func,
 };
+function mapStateToProps(state) {
+  const {
+    modal, map, measure, vectorStyles, browser, compare, proj, ui, layers,
+  } = state;
+  let swipeOffset;
+  const activeLayers = layers[compare.activeString];
+  if (compare.active && compare.mode === 'swipe') {
+    const percentOffset = state.compare.value || 50;
+    swipeOffset = browser.screenWidth * (percentOffset / 100);
+  }
+
+  return {
+    modalState: modal,
+    isShowingClick: map.isClickable,
+    isDistractionFreeModeActive: ui.isDistractionFreeModeActive,
+    getDialogObject: (pixels, olMap) => onMapClickGetVectorFeatures(pixels, olMap, state, swipeOffset),
+    lastSelected: vectorStyles.selected,
+    measureIsActive: measure.isActive,
+    isMobile: browser.lessThan.medium,
+    compareState: compare,
+    swipeOffset,
+    proj,
+    activeLayers,
+  };
+} const mapDispatchToProps = (dispatch) => ({
+  selectVectorFeatures: (features) => {
+    setTimeout(() => {
+      dispatch(selectVectorFeaturesActionCreator(features));
+    }, 1);
+  },
+  changeCursor: (bool) => {
+    dispatch(changeCursorActionCreator(bool));
+  },
+  openCustomAlertModal: ({ id, props }) => {
+    dispatch(openCustomContent(id, props));
+  },
+  onCloseModal: () => {
+    dispatch(onClose());
+  },
+  activateVectorAlert: () => dispatch({ type: ACTIVATE_VECTOR_ALERT }),
+  openVectorDiaglog: (dialogId, metaArray, offsetLeft, offsetTop, isMobile) => {
+    const dialogKey = new Date().getUTCMilliseconds();
+    dispatch(openCustomContent(dialogId,
+      {
+        backdrop: false,
+        clickableBehindModal: true,
+        desktopOnly: true,
+        isDraggable: true,
+        wrapClassName: 'vector-modal-wrap',
+        modalClassName: 'vector-modal light',
+        CompletelyCustomModal: vectorDialog,
+        isResizable: true,
+        dragHandle: '.modal-header',
+        dialogKey,
+        key: dialogKey,
+        vectorMetaObject: lodashGroupBy(metaArray, 'id'),
+        width: isMobile ? 250 : 445,
+        height: 300,
+        offsetLeft,
+        offsetTop,
+        timeout: 0,
+        onClose: () => {
+          setTimeout(() => {
+            dispatch(selectVectorFeaturesActionCreator({}));
+          }, 1);
+        },
+      }));
+  },
+});
+VectorInteractions.propTypes = {
+  changeCursor: PropTypes.func.isRequired,
+  getDialogObject: PropTypes.func.isRequired,
+  isShowingClick: PropTypes.bool.isRequired,
+  measureIsActive: PropTypes.bool.isRequired,
+  modalState: PropTypes.object.isRequired,
+  mouseEvents: PropTypes.object.isRequired,
+  onCloseModal: PropTypes.func.isRequired,
+  openVectorDiaglog: PropTypes.func.isRequired,
+  selectVectorFeatures: PropTypes.func.isRequired,
+  compareState: PropTypes.object,
+  isMobile: PropTypes.bool,
+  lastSelected: PropTypes.object,
+  proj: PropTypes.object,
+  swipeOffset: PropTypes.number,
+  activeLayers: PropTypes.array,
+  activateVectorAlert: PropTypes.func,
+};
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(VectorInteractions);
