+++ web/js/containers/map-interactions/ol-vector-interactions.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import util from '../../util/util';
-import VectorInteractions from './ol-vector-interactions';
+import { VectorInteractions } from './ol-vector-interactions';
 import { registerProjections } from '../../fixtures';
 
 let events;
